Fix warning

+++ src/server/game/Entities/Unit/Unit.cpp
@@ -17332,7 +17332,7 @@ uint32 Unit::GetRemainingDotDamage(uint64 caster, uint32 spellId, uint8 effectIn
     {
         if ((*i)->GetCasterGUID() != caster || (*i)->GetId() != spellId || (*i)->GetEffIndex() != effectIndex)
             continue;
-        if ((*i)->GetTotalTicks() >= (*i)->GetTickNumber())
+        if ((*i)->GetTotalTicks() >= int32((*i)->GetTickNumber()))
             amount += ((*i)->GetAmount() * ((*i)->GetTotalTicks() - ((*i)->GetTickNumber()))) / (*i)->GetTotalTicks();
         break;
     }
